feat(build): migrate from travis to GitHub actions - #235 (#316)
* feature(build) : Migrate from Travis CI to GitHub actions - 235


* chore: Fixing global testing and peer dependencies (#317)


* fix(package-lock): Update package-lock.json to fix tests


* feat(build):  Update build step and publish step inside github workflows


* feat(build): Remove -notify dependency from publish job


* feat(build): Add cross-env as dev dependency for windows build
