+++ .github/workflows/build.yml
@@ -1,0 +1,71 @@
+name: Build
+
+on:
+  push:
+  pull_request:
+    branches:
+      - master
+
+jobs:
+  build:
+    name: Unit Tests
+
+    strategy:
+      matrix:
+        node-version:
+          - 12.x
+          - 14.x
+        os:
+          - ubuntu-latest
+          - windows-latest
+          - macOS-latest
+
+    runs-on: ${{ matrix.os }}
+
+    steps:
+      - name: git checkout
+        uses: actions/checkout@v2
+
+      - name: Use Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@v1
+        with:
+          node-version: ${{ matrix.node-version }}
+    
+      - run: npx lerna bootstrap && npx lerna run build
+      - run: npm run build --if-present
+      - run: npm test
+
+  publish:
+    needs:
+      - build
+
+    name: Publish to npm
+    if: ${{ success() && github.event_name == 'push' && github.repository_owner == 'accordproject' }}
+    runs-on: ubuntu-latest
+
+    outputs:
+      job-status: ${{ job.status }}
+
+    steps:
+      - name: git checkout
+        uses: actions/checkout@v2
+
+      - name: Use Node.js 14.x
+        uses: actions/setup-node@v1
+        with:
+          node-version: 14.x
+    
+      - name: Build
+        run: npx lerna bootstrap && npx lerna run build
+
+      - name: timestamp
+        id: timestamp
+        run: |
+          node ./scripts/timestamp.js
+
+      - name: build and publish
+        run: |
+          npm config set //registry.npmjs.org/:_authToken ${{ secrets.NPM_TOKEN }}
+          npm version --no-git-tag-version --yes --exact ${{ steps.timestamp.outputs.stamp }}
+          npx lerna version --no-git-tag-version --yes --exact ${{ steps.timestamp.outputs.stamp }}
+          npx lerna exec --ignore storybook -- npm publish --access public --ignore-scripts --tag=unstable 2>&1
