--- .travis.yml
@@ -1,29 +1,0 @@
-language: node_js
-node_js:
-  - '10'
-dist: trusty
-before_install: |
-  set -e
-  export PATH="./$HOME/node_modules/.bin:$PATH"
-  npm install -g @alrra/travis-scripts
-  date
-install: |
-  set -e
-  date
-  npm install -g lerna@^3.15.0
-  lerna bootstrap 2>&1
-  date
-script: lerna run build:prod
-deploy:
-  provider: script
-  script: 'if [ "$TRAVIS_PULL_REQUEST" = "false" ]; then bash ./.travis/deploy.sh; fi'
-  skip_cleanup: true
-  on:
-    all_branches: true
-    repo: accordproject/web-components
-    condition: "$TRAVIS_OS_NAME = linux"
-after_failure: tail -n +1 -- /home/travis/.npm/_logs/*-debug.log
-after_success: 
-cache:
-  directories:
-  - node_modules
