+++ README.md
@@ -1,6 +1,6 @@
 # Accord Project Web Components
 
-[![Build Status](https://travis-ci.org/accordproject/web-components.svg?branch=master)](https://travis-ci.org/accordproject/web-components) 
+[![Build Status](https://github.com/accordproject/web-components/actions/workflows/build.yml/badge.svg)](https://github.com/accordproject/web-components/actions/workflows/build.yml)
 [![join slack](https://img.shields.io/badge/Accord%20Project-Join%20Slack-blue)](https://accord-project-slack-signup.herokuapp.com/)
 
 This repository contains a library of user interface components that can be used to create web-applications based on Accord Project technology. The components use the [React](https://reactjs.org) web application framework.
