+++ package.json
@@ -34,19 +34,13 @@
   },
   "scripts": {
     "storybook": "lerna exec --scope storybook -- npm run build-storybook",
-    "postinstall": "npm run pkgcheck && npm run bootstrap",
+    "postinstall": "npm run bootstrap",
     "bootstrap": "lerna bootstrap",
     "licchk": "license-check-and-add",
     "pretest": "npm run licchk",
     "test:e2e": "lerna exec --scope storybook -- npm run ci:start-and-test-storybook",
     "test": "npm run test:e2e",
-    "publish": "./scripts/manualrelease.sh",
     "repoclean": "lerna clean",
-    "pkgcheck": "node ./scripts/pkgcheck.js",
-    "pkgstamp": "node ./scripts/pkgstamp.js",
-    "pkgbump": "node ./scripts/pkgbump.js && node ./scripts/pkgcheck.js --fix",
-    "pkgset": "node ./scripts/pkgset.js",
-    "depcheck": "node ./scripts/depcheck.js",
     "changelog:unreleased": "lerna-changelog"
   },
   "repository": {
